tests: Add a self-hosted test for libwalbottle
This uses the meta-schemas to generate test vectors which are valid and
invalid JSON schemas. Each of those test vectors is parsed as a
WblSchema and used to generate JSON instances. The instances themselves
are not tested, but the correspondence between
WblGeneratedInstance.valid and success in parsing and generating using
each test vector is checked.

This highlights a number of areas where Walbottle will generate a JSON
schema from the meta-schema and mark it as valid (for example), but will
then refuse to parse it as it’s actually invalid. This comes from a few
key as-yet-unimplemented features in Walbottle, such as $ref
(json-schema-core§7) and $schema (json-schema-core§6).

These failures are tracked as expected failures in self-hosting.c, and
hopefully will decrease as appropriate featurese are added.

By itself, the self-hosting test achieves 75.6% line coverage of
libwalbottle (74.6% line coverage of wbl-schema.c); versus 91.1% (91.3%)
achieved by the schema and schema-keywords tests. There’s still a bit of
work to do before self-hosting is a complete success.