lib: Re-work the hack for generating instances of regex patterns
The old hack could fall into an infinite loop if a simple string could
not be generated to match the pattern. Replace it with a different hack
which chooses one of a fixed number of strings, but fails with a clear
error message if none of those match.

The ultimate solution is to write something which will work back through
the FSM for the pattern, but that requires a significant effort.