lib: Generate instances for all subschemas of schema array keywords
Generation of instances for keywords:
 • oneOf
 • anyOf
 • allOf
used to be controlled by some predicates which chose only some of the
instances generated for the subschemas in the keyword’s array. This
doesn’t make sense from a code coverage point of view, as it’s
eliminating legitimate instances which could exercise code belonging to
one of those subschemas (even if those instances ultimately don’t match
the right number of the subschemas to be valid).

Instead of trying complicated things like that, just emit all the
generated instances for the subschemas.