lib: Further redesign of instance generation for array keywords
Following on from commit 573c18d6ea859498dc8c70f1bfc935c8d1f83eb5,
introduce the concept of validity arrays to partially formalise the
relationship between generated array instances and the control flows in
the code under test.

A validity array is a boolean array which controls which child elements
of a generated array instance are generated as valid or as invalid,
according to the relevant subschema. By separating validity array
generation from child instance generation, the problems of:
 • generating array instances appropriate for testing different control
   flows; and
 • generating enough array instances to include all generated child
   instances in at least one valid array instance
are simpler to solve.