lib: Add a second method for comparing object member name lists
As it’s called from node_equal() to compare two objects for most hash
operations, member_names_equal() is a very hot path in the code. For
small lists, it’s actually a lot faster to do the O(N^2) direct list
comparison than it is to allocate a hash table and do all the string
hashing — so implement that.

This reduces `time make check` from 1:06 minutes to 41 seconds.