lib: Hoist some expensive loop-invariant calculations
The code was calling subschema_generate_instances() and validating each
of them multiple times for the same subschema as it appeared in
different subschema arrays. Avoid that by hoisting the calculation into
a loop-invariant map from subschemas to their valid and invalid
instances.

This reduces `time make check` from 3:00 to 2:40 minutes.