lib: Implement a cache of instances generated for each subschema
It turns out that despite various levels of caching in
generate_all_properties() and generate_all_items(), the instances for a
given subschema in a multi-level schema document were still being
regenerated a super-linear number of times during the generation
process.

This doesn’t affect `time make check` (I believe the majority of that
time is now spent in parsing and comparing the results), but does speed
up a pathological nested subschema case from ~16 minutes to 32 seconds.