lib: Implement object generation using validity objects
This separates out the policy of which generated object instance members
are valid compared to the set of valid property sets. This makes it
easier to ensure that all valid subinstances are included in an
overall-valid object instance, so that deep code paths which are only
accessible when the entire JSON test vector is valid are reached.

This is similar to the previous work on using validity arrays for JSON
array generation.