utilities: Redirect all debug messages to stderr rather than stdout
This means that all GLog output goes to stderr, and all useful output
(like generated test vectors) goes to stdout.

This means that you can safely redirect the stdout of
json-schema-generate to a test vector file even when G_MESSAGES_DEBUG is
defined — this no longer results in a test vector file full of debug
messages.