lib: Expand error reporting for schema validation
Instead of reporting a single error for the entirety of the validation
process of a schema, return an array of errors, covering all things the
validator was able to check (i.e. proceeding past the first error).
Errors are returned in a new WblValidateMessage struct, which contains
more information than a GError, such as the relevant specification
section, and the JsonNode which caused the problem.

The error returned by wbl_schema_load_from_stream() and friends is now a
generic error which is set if any errors are returned by the validator.

This breaks ABI by changing the definition of WblSchemaClass. It also
adds new API:
 • wbl_schema_get_validate_messages()
 • WblValidateMessageLevel
 • WblValidateMessage
 • wbl_validate_message_copy()
 • wbl_validate_message_free()
 • WBL_SCHEMA_CORE
 • WBL_SCHEMA_VALIDATION