lib: Improve floating point remainder checks for multipleOf
On some i386 platforms, apparently the error introduced by casting a
gint64 instance value to a gdouble was big enough to cause a non-zero
remainder when dividing the resulting gdouble by a gdouble which should
divide it.

For example, this happened with instance value 12 and schema value 1.2.

Try and avoid this by considering the errors introduced by the cast in
the checks, and comparing against suitable multiples of machine epsilon
to ignore these errors.

Note that this is tested against the test suite on a 64-bit machine, but
not on the errant i386 machine because I do not have access to it.