lib: Combine two floating point paths
Since json-glib has potentially introduced an error when converting the
string representation of an integeric or floating point instance or
schema value to a double, we always have to account for both the error
in i and s, so we might as well combine the code paths.

Add a negative test to check that the next floating point number after
the factor doesn’t divide.